Fix crashes on files around 2 GiB in size
quick-lint-js invokes undefined behavior (signed integer overflow) in a
few spots when given a file around 2 GiB in size (2147483632 to
2147483647 bytes, specifically). Fix this UB to prevent crashes with
UBSAN and possibly future bugs.

+++ src/quick-lint-js/padded-string.h
@@ -40,7 +40,8 @@ class padded_string {
   char8 *data() noexcept { return this->data_.data(); }
 
   int size() const noexcept {
-    return narrow_cast<int>(this->data_.size()) - this->null_bytes_to_add;
+    return narrow_cast<int>(this->data_.size() -
+                            narrow_cast<unsigned>(this->null_bytes_to_add));
   }
 
   const char8 &operator[](int index) const noexcept {
